删除学生，删除考核轮，删除考核周期完成

+++ src/main/java/com/jacky/register/server/localFiles/StorageService.java
@@ -89,6 +89,7 @@ public interface StorageService<Model> {
         }
     }
     static void delete(String path,Path savePath) {
+        if(path==null)return;
         File file = savePath.resolve(path).toFile();
 
         FileSystemUtils.deleteRecursively(file);
